PyHardLinkBackup/requirements/dev_extras.txt

# convert README.creole on-the-fly to ReSt, see also:
# https://github.com/jedie/python-creole/wiki/Use-In-Setup/
python-creole
docutils

# Build and send to PyPi:
wheel
twine

# unittests
nose
django-nose
coveralls